Refactor inspector controls into helpers
