+++ src/frontend.js
@@ -1,4 +1,4 @@
-import lazyLoadVideos from './assets/js/lazyVideos'
+import lazyLoadVideos from './assets/js/lazyVideos';
 import initCustomPlayButtons from './assets/js/customPlayButtons';
 
 lazyLoadVideos();
